+++ test/unit/integrations/mongoid_test.rb
@@ -1567,6 +1567,16 @@ module MongoidTest
       assert_equal 'shutdown', machine.state(:parked).human_name
     end
     
+    def test_should_support_nil_state_key
+      I18n.backend.store_translations(:en, {
+        :mongoid => {:state_machines => {:states => {:nil => 'empty'}}}
+      })
+      
+      machine = StateMachine::Machine.new(@model)
+      
+      assert_equal 'empty', machine.state(nil).human_name
+    end
+    
     def test_should_allow_customized_event_key_scoped_to_class_and_machine
       I18n.backend.store_translations(:en, {
         :mongoid => {:state_machines => {:'mongoid_test/foo' => {:state => {:events => {:park => 'stop'}}}}}